+++ vgo-core/src/main/kotlin/com/jzbrooks/vgo/core/graphic/ElementPrinter.kt
@@ -3,10 +3,10 @@ package com.jzbrooks.vgo.core.graphic
 /**
  * Serializes geometry elements in a target document format.
  *
- * Implementations print geometry only. Paint attributes are equal
- * across the representations these strings are used to compare.
+ * Implementations print geometry only.
+ * Paint attributes are ignored in outputs.
  */
-interface ShapePrinter {
+interface ElementPrinter {
     fun print(shape: Shape): String
 
     fun print(path: Path): String
