+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/graphic/FakeElementPrinter.kt
@@ -1,19 +1,20 @@
-package com.jzbrooks.vgo.svg
+package com.jzbrooks.vgo.core.graphic
 
-import com.jzbrooks.vgo.core.graphic.Circle
-import com.jzbrooks.vgo.core.graphic.Ellipse
-import com.jzbrooks.vgo.core.graphic.Line
-import com.jzbrooks.vgo.core.graphic.Path
-import com.jzbrooks.vgo.core.graphic.Polygon
-import com.jzbrooks.vgo.core.graphic.Polyline
-import com.jzbrooks.vgo.core.graphic.Rect
-import com.jzbrooks.vgo.core.graphic.Shape
-import com.jzbrooks.vgo.core.graphic.ShapePrinter
+import com.jzbrooks.vgo.core.graphic.command.FakeCommandPrinter
 import com.jzbrooks.vgo.core.util.math.Point
+import java.math.RoundingMode
+import java.text.DecimalFormat
 
-class ScalableVectorGraphicShapePrinter(
-    private val commandPrinter: ScalableVectorGraphicCommandPrinter,
-) : ShapePrinter {
+class FakeElementPrinter : ElementPrinter {
+    private val commandPrinter = FakeCommandPrinter()
+
+    private val formatter =
+        DecimalFormat().apply {
+            maximumFractionDigits = 2
+            isDecimalSeparatorAlwaysShown = false
+            roundingMode = RoundingMode.HALF_UP
+        }
+
     override fun print(path: Path): String =
         buildString {
             append("<path d=\"")
@@ -33,13 +34,8 @@ class ScalableVectorGraphicShapePrinter(
             }
 
             is Rect -> {
-                buildString {
-                    append("<rect x=\"${print(shape.x)}\" y=\"${print(shape.y)}\" ")
-                    append("width=\"${print(shape.width)}\" height=\"${print(shape.height)}\"")
-                    if (shape.rx > 0f) append(" rx=\"${print(shape.rx)}\"")
-                    if (shape.ry > 0f) append(" ry=\"${print(shape.ry)}\"")
-                    append("/>")
-                }
+                "<rect x=\"${print(shape.x)}\" y=\"${print(shape.y)}\" " +
+                    "width=\"${print(shape.width)}\" height=\"${print(shape.height)}\"/>"
             }
 
             is Line -> {
@@ -56,7 +52,7 @@ class ScalableVectorGraphicShapePrinter(
             }
         }
 
-    private fun print(value: Float) = commandPrinter.formatter.format(value)
+    private fun print(value: Float) = formatter.format(value)
 
     private fun print(points: List<Point>) = points.joinToString(" ") { "${print(it.x)},${print(it.y)}" }
 }
